First Screen Recording
I've added a Main Menu, of sorts, a character image, and a few text boxes that transition from one to the other when the "Next" button is clicked. I've also made it so that clicking the "Next" button when there are no more text boxes to be loaded takes you to the next scene (which I've set to be the Main Menu page for now since I don't have any other scenes).