Fixing Starting Scene, Starting Second Level
Fixed the starting scene so that the Next Text Box Button and Text Box Changer have their own scripts rather than using the same scripts as their counterpart in the level scenes. This was done since there were elements of those scenes that needed to be added to the Button and Text Box Changer that did not exist, or need to exist, in the starting scene. I also started creating the second level, and added a new graphic, a rabbit, for this level.