Started creating level 2 (technically level 1)
I started creating the basic framework and layout for level 2 (technically level 1 since I started with level 0) of my game. This was mostly the same as the framework of the first level (level 0), although I added a rabbit, a button to click the rabbit, and a panel that pops up when the rabbit is clicked so that the rabbit can be fed, amongst other things.