Implementing New Level Button Disabling
Started implementing a new version of the level button disabling. What this is doing is that, in the Map scene, where I have all of my level buttons, the level buttons will only take players to the level if they have completed all other levels. This works by updating the value of the last level passed every time a level is passed, checking the level number of the button clicked, and only letting the player access the scene for a level if the level number of the last level passed is greater than or equal to the level number of the button clicked. Also fixed a bug where the buttons for the levels after the first (technically 0th) level would not work.