Implementing New Button Disabling Method
In the process of creating a new method to track how many levels a player has passed, and only letting them access a level after all of the previous ones have been passed. This works by having a counter for the last level passed, adding a level value to each level button, comparing the counter and the level value, and only letting the player access the level if the value of the counter is greater than or equal to the level value of the button they clicked. Also fixed a bug where none of the level buttons other than the one for the first (technically 0th) level was working.