Improved Dialogue, Fixing Small Bugs, Adding More Dialogue
Improved the dialogue system to effectively show an image of the character speaking in the bottom left corner. Fixed small bugs, such as going directly to the next part of the scene after the rabbit is fed, only letting the rabbit be fed once, and changing the dialogue appearance slightly. Finished the dialogue for Level 1 (the second level).