Updated Rabbit Feeding, Fixed a Bug
Updated rabbit feeding to only let the rabbit be fed if the player has enough berries, and otherwise telling them they don't have enough berries and sending them back to the map to get more. Also fixed a bug that kept giving players berries infinitely in the first level since it was the last text box in the level.