Fixed Berry Counter Display, Fixed Dialogue Bug
Fixed the berry counter display (image + counter/number of berries) so that it renders above the background, but below panels. Also fixed the panels and dialogue so that the panels render on top of everything instead of the dialogue.  Also played around with the placement of the berry counter within the scene to find a good position for it.