Adding Decision-Making
Started adding decision-making functionality. This involved creating a decision panel, buttons for the options available to choose from, and a way to record these decisions to determine which version of the next level to show (I'm using my LevelCounter function to store these decisions so I can integrate them into the Map's level buttons). So far, I have most of the code, and a basic layout of the decision panel, but I need to debug since clicking the decision buttons currently gives me an IndexOutOfRangeException.