Level 2.1-1 Debugging
Debugged the dialogue so that the rabbit can be given berries without the text in the text box saying to do so showing in the dialogue. Also found an Index Out Of Range error appears when I try to use the decision making panel buttons, so I need to fix that.