Updated Map Level Buttons
Started updating map level button OnClick() functions to include which version of the level to show based on the decisions made in the past. I started creating a separate class for this, but then decided to start using the pre-existing function LevelButtonOnClick() in Level Counter instead so it's less convoluted.