Added the scenes and buttons for Level 3
Added the button for Level 3 in the Map, as well as the scenes for the level. I've finished the first part of the dialogue and the decision for the level, so now I just need to make the dialogues for the second part of the level, after a decision is made.