Added Level 3-1
Added Level 3-1, which is what happens after the player chooses the second option in Level 2's decision, as well as a decision in Level 3-1 and the two scenarios it can take the player to. I've made the dialogue and scenes for them, but I haven't integrated them into the rest of the game yet, so Level 2 always goes to Level 3-0. I'll be fixing this later.