Added Level Labels and End Panel; Added a Bug
Added level number labels under the level buttons in the map to make it clearer for players what they're supposed to click. I also added an end scene/panel that appears when the player has reached the end of a story line. However, I realized after that I managed to add a bug, which is that Level 2 shows the berries acquired panel upon opening it, and it won't close when the "OK" button is clicked, so I'm trying to fix that now.